Move fetch stub helpers to test-utils and fix coverage gap
Moved useFetchStub() from the underscore-prefixed _fetch-helpers.ts into
src/test-utils/index.ts to match the codebase convention where all shared
test helpers live in #test-utils. No other test subdirectory has a local
helper file.

Inlined error spy usage directly in send-email.test.ts and config.test.ts
using spy(console, "error") + try/finally, matching the pattern used
across the rest of the test suite. This eliminates the custom
expectEmailSendLog helper whose error-throw branch was unreachable,
fixing the coverage gap.
